+++ examples/surreal_ql/ql_show_examples.py
@@ -12,10 +12,10 @@ with Database("http://127.0.0.1:8000", 'test', 'test', credentials=('user_db', '
 
     tm = to_surreal_datetime_str(datetime.now(timezone.utc))  # get current time in surreal format
 
-    # SHOW CHANGES FOR TABLE person SINCE "2023-09-07T01:23:52Z";
+    # SHOW CHANGES FOR TABLE person SINCE d"2023-09-07T01:23:52Z";
     print(db.table("person").show_changes().since(tm))
     print(db.table("person").show_changes(since=tm))
 
-    # SHOW CHANGES FOR TABLE person SINCE "2023-09-07T01:23:52Z" LIMIT 10;
-    print(db.person.show_changes().since("2023-09-07T01:23:52Z").limit(10))
-    print(db.person.show_changes(since="2023-09-07T01:23:52Z").limit(10))
+    # SHOW CHANGES FOR TABLE person SINCE "d'2023-09-07T01:23:52Z'" LIMIT 10;
+    print(db.person.show_changes().since("d'2023-09-07T01:23:52Z'").limit(10))
+    print(db.person.show_changes(since="d'2023-09-07T01:23:52Z'").limit(10))