+++ src/surrealist/ql/statements/show.py
@@ -1,9 +1,8 @@
-import json
 from datetime import datetime, timezone
 from typing import List, Optional
 
 from surrealist.connections import Connection
-from surrealist.utils import OK, DATE_FORMAT, DATE_FORMAT_NS, to_surreal_datetime_str
+from surrealist.utils import OK, to_surreal_datetime_str
 from .statement import Statement
 
 
@@ -28,11 +27,8 @@ class Show(Statement):
     def validate(self) -> List[str]:
         result = []
         if self._since and isinstance(self._since, str):
-            try:
-                format_ = DATE_FORMAT if "." not in self._since else DATE_FORMAT_NS
-                datetime.strptime(self._since, format_)
-            except ValueError:
-                result.append("Timestamp in the wrong format, you need iso-date like 2024-01-01T10:10:10.000001Z")
+            if not self._since.startswith("d"):
+                result.append("Timestamp in the wrong format, you need iso-date like d'2024-01-01T10:10:10.000001Z'")
         if self._limit and self._limit < 1:
             result.append("Limit should not be less than 1")
         return [OK] if not result else result
@@ -48,7 +44,7 @@ class Show(Statement):
 
     def since(self, timestamp: str) -> "Show":
         """
-        Init timestamp since is to show updates
+        Init timestamp since is to show updates, it should be a surreal timestamp like d'2024-01-01T10:10:10.000001Z'
 
         Refer to: https://surrealdb.com/docs/surrealdb/surrealql/statements/show#basic-usage
 
@@ -62,4 +58,4 @@ class Show(Statement):
         if not self._since:
             self._since = to_surreal_datetime_str(datetime.now(timezone.utc))  # default value
         limit = f" LIMIT {self._limit}" if self._limit else ""
-        return f'SHOW CHANGES FOR TABLE {self._table_name} SINCE d{json.dumps(self._since)}{limit}'
+        return f'SHOW CHANGES FOR TABLE {self._table_name} SINCE {self._since}{limit}'