+++ tests/unit_tests/test_utils.py
@@ -1,15 +1,17 @@
 import datetime
 from unittest import TestCase, main
 
-from surrealist.utils import to_datetime, to_surreal_datetime_str, crop_data, mask_pass
+from surrealist.utils import to_datetime, to_surreal_datetime_str, crop_data, mask_pass, clean_dates
 
 
 class TestUtils(TestCase):
     def test_to_datetime(self):
         self.assertEqual(to_datetime('2018-01-01T00:00:00.000000Z'), datetime.datetime(2018, 1, 1, 0, 0))
+        self.assertEqual(to_datetime('d"2018-01-01T00:00:00.000000Z"'), datetime.datetime(2018, 1, 1, 0, 0))
+        self.assertEqual(to_datetime("d'2018-01-01T00:00:00.000000Z'"), datetime.datetime(2018, 1, 1, 0, 0))
 
     def test_to_surreal_datetime_str(self):
-        self.assertEqual(to_surreal_datetime_str(datetime.datetime(2018, 1, 1, 0, 0)), '2018-01-01T00:00:00.000000Z')
+        self.assertEqual(to_surreal_datetime_str(datetime.datetime(2018, 1, 1, 0, 0)), "d'2018-01-01T00:00:00.000000Z'")
 
     def test_crop_same(self):
         self.assertEqual(crop_data("one"), "one")
@@ -22,6 +24,18 @@ class TestUtils(TestCase):
             "{'method': 'signin', 'params': [{'user': 'root', 'pass': '******', 'NS': 'test', 'DB': 'test'}]}"),
             "{'method': 'signin', 'params': [{'user': 'root', 'pass': '******', 'NS': 'test', 'DB': 'test'}]}")
         self.assertEqual(mask_pass('{"user":"user", "pass": "123123"}'), '{"user":"user", "pass": "******"}')
+
+    def test_clean_dates(self):
+        text = ""
+        self.assertEqual(clean_dates(text), text)
+        text = "2018-01-01T00:00:00.000000Z"
+        self.assertEqual(clean_dates(text), text)
+        text = """CREATE z CONTENT {"name": "xxx", "age": 22, "create_time": "d'2024-10-23T16:06:51.322496Z'"};"""
+        expected = """CREATE z CONTENT {"name": "xxx", "age": 22, "create_time": d'2024-10-23T16:06:51.322496Z'};"""
+        self.assertEqual(clean_dates(text), expected)
+        text = """CREATE z CONTENT {"name": "xxx", "age": 22, "create_time": 'd"2024-10-23T16:06:51.322496Z"'};"""
+        expected = """CREATE z CONTENT {"name": "xxx", "age": 22, "create_time": d"2024-10-23T16:06:51.322496Z"};"""
+        self.assertEqual(clean_dates(text), expected)
 
 
 if __name__ == '__main__':